backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
python-multipart
pyautogui
selenium
python-dotenv
chromadb
pytesseract
pillow
requests
